[Instrumentor] A configurable instrumentation pass
Both internally and externally we instrument code all the time. Usually,
those are hand written traversals that look for some specific
instructions or program points and then provide information to a
runtime. Since the real difference is in the runtime, and the
instrumentation is basically always the same, people can use this
Instrumentor pass and configure it to their needs.

Initial implementation only instruments alloca instructions but shows
the setup, and configurability via a JSON file.